all pick ups
Il faudrat modifier les pick ups de dmgUp, berserk et epines par rapport à l'interface de dégâts.

(interface de dégâts qui fait que je devrais modifier l'herbe aussi)

+++ Assets/IA/CS_Enemy.cs
@@ -16,7 +16,7 @@ public class CS_Enemy : MonoBehaviour
     protected Transform playerTransform;
     protected Animator animator;
 
-    [SerializeField] protected int PV = 3;
+    [SerializeField] protected float PV = 3;
     [SerializeField] protected GameObject prefab_DeathFX;
 
     [SerializeField] Material stunningMaterial;
@@ -80,7 +80,7 @@ public class CS_Enemy : MonoBehaviour
         timeStopStunning = Time.time + duration;
     }
 
-    virtual public void TakeDamage(int damage, PlayerAttackType type)
+    virtual public void TakeDamage(float damage, PlayerAttackType type)
     {
         touched = true;
         PV -= damage;
